+++ CHANGELOG.md
@@ -13,7 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 * Add `egui::plot::Plot` to plot some 2D data.
 * Add `Ui::hyperlink_to(label, url)`.
-* Sliders can now have a value prefix and suffix (e.g. "°" as a unit).
+* Sliders can now have a value prefix and suffix (e.g. the suffix `"°"` works like a unit).
+* `Context::set_pixels_per_point` to control the scale of the UI.
 
 ### Changed 🔧